StockSizingAI/requirements.txt
fastapi
uvicorn
numpy
pandas
yfinance
matplotlib
riskfolio-lib
pydantic
stripe
clerk-backend-api
